kahve-zeka-backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
# passlib[bcrypt] SATIRI SİLİNDİ, YERİNE BUNLAR GELDİ:
passlib
bcrypt<4.0
# -------------------------------------
python-jose[cryptography]
python-multipart
psycopg2-binary
python-dotenv